BE/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
openai
pydantic
ultralytics
torch 
torchvision